LayoutTests/platform/chromium-mac/fast/js/console-non-string-values-expected.txt

CONSOLE MESSAGE: line 12: I am a string
CONSOLE MESSAGE: line 13: 42
Should see two console messages, the first a string, and the second a number.
PASS successfullyParsed is true

TEST COMPLETE
